move WaitForTidbClusterReady out of operatorActions
there are still a lot of operator unrelated functions needs to be
moved out from operatorActions. At the last, we will have a clean
struct and no redundant interface.
