+++ tests/actions.go
@@ -37,7 +37,6 @@ import (
 	"github.com/pingcap/advanced-statefulset/client/apis/apps/v1/helper"
 	asclientset "github.com/pingcap/advanced-statefulset/client/client/clientset/versioned"
 	pingcapErrors "github.com/pingcap/errors"
-	"github.com/pingcap/kvproto/pkg/metapb"
 	"github.com/pingcap/tidb-operator/pkg/apis/pingcap/v1alpha1"
 	"github.com/pingcap/tidb-operator/pkg/client/clientset/versioned"
 	"github.com/pingcap/tidb-operator/pkg/controller"
@@ -58,7 +57,6 @@ import (
 	"github.com/pingcap/tidb-operator/tests/pkg/webhook"
 	"github.com/pingcap/tidb-operator/tests/slack"
 	admissionV1beta1 "k8s.io/api/admissionregistration/v1beta1"
-	apps "k8s.io/api/apps/v1"
 	v1 "k8s.io/api/apps/v1"
 	batchv1 "k8s.io/api/batch/v1"
 	corev1 "k8s.io/api/core/v1"
@@ -74,6 +72,7 @@ import (
 	"k8s.io/kubernetes/test/e2e/framework"
 	"k8s.io/kubernetes/test/e2e/framework/log"
 	"k8s.io/kubernetes/test/e2e/framework/pod"
+	ctrlCli "sigs.k8s.io/controller-runtime/pkg/client"
 )
 
 const (
@@ -89,11 +88,13 @@ const (
 	WebhookServiceName = "webhook-service"
 )
 
-func NewOperatorActions(cli versioned.Interface,
+func NewOperatorActions(
+	cli versioned.Interface,
 	kubeCli kubernetes.Interface,
 	asCli asclientset.Interface,
 	aggrCli aggregatorclientset.Interface,
 	apiExtCli apiextensionsclientset.Interface,
+	genericCli ctrlCli.Client,
 	pollInterval time.Duration,
 	operatorConfig *OperatorConfig,
 	cfg *Config,
@@ -119,7 +120,7 @@ func NewOperatorActions(cli versioned.Interface,
 		pollInterval: pollInterval,
 		cfg:          cfg,
 		fw:           fw,
-		crdUtil:      NewCrdTestUtil(cli, kubeCli, asCli, tcStsGetter),
+		crdUtil:      NewCrdTestUtil(cli, kubeCli, asCli, genericCli, nil),
 	}
 	if fw != nil {
 		kubeCfg, err := framework.LoadConfig()
@@ -165,10 +166,6 @@ type OperatorActions interface {
 	DumpAllLogs(info *OperatorConfig, clusterInfos []*TidbClusterConfig) error
 	DeployTidbCluster(info *TidbClusterConfig) error
 	DeployTidbClusterOrDie(info *TidbClusterConfig)
-	CleanTidbCluster(info *TidbClusterConfig) error
-	CleanTidbClusterOrDie(info *TidbClusterConfig)
-	CheckTidbClusterStatus(info *TidbClusterConfig) error
-	CheckTidbClusterStatusOrDie(info *TidbClusterConfig)
 	BeginInsertDataTo(info *TidbClusterConfig) error
 	BeginInsertDataToOrDie(info *TidbClusterConfig)
 	StopInsertDataTo(info *TidbClusterConfig)
@@ -192,7 +189,6 @@ type OperatorActions interface {
 	Restore(from *TidbClusterConfig, to *TidbClusterConfig) error
 	CheckRestore(from *TidbClusterConfig, to *TidbClusterConfig) error
 	RestoreIncrementalFiles(from *DrainerConfig, to *TidbClusterConfig, stopTSO int64) error
-	ForceDeploy(info *TidbClusterConfig) error
 	CreateSecret(info *TidbClusterConfig) error
 	GetPodUIDMap(info *TidbClusterConfig) (map[string]types.UID, error)
 	GetNodeMap(info *TidbClusterConfig, component string) (map[string][]string, error)
@@ -238,7 +234,6 @@ type OperatorActions interface {
 	CheckInitSQL(info *TidbClusterConfig) error
 	CheckInitSQLOrDie(info *TidbClusterConfig)
 	DeployAndCheckPump(tc *TidbClusterConfig) error
-	WaitForTidbClusterReady(tc *v1alpha1.TidbCluster, timeout, pollInterval time.Duration) error
 	WaitPodOnNodeReadyOrDie(clusters []*TidbClusterConfig, faultNode string)
 	DataIsTheSameAs(from, to *TidbClusterConfig) (bool, error)
 }
@@ -811,299 +806,6 @@ func (oa *operatorActions) DeployTidbClusterOrDie(info *TidbClusterConfig) {
 	}
 }
 
-func (oa *operatorActions) CleanTidbCluster(info *TidbClusterConfig) error {
-	log.Logf("cleaning tidbcluster %s/%s", info.Namespace, info.ClusterName)
-	oa.EmitEvent(info, "CleanTidbCluster")
-	ns := info.Namespace
-	tcName := info.ClusterName
-
-	oa.StopInsertDataTo(info)
-
-	selector, err := metav1.LabelSelectorAsSelector(&metav1.LabelSelector{
-		MatchLabels: map[string]string{
-			label.InstanceLabelKey: tcName,
-		},
-		MatchExpressions: []metav1.LabelSelectorRequirement{
-			{
-				Key:      label.ComponentLabelKey,
-				Operator: metav1.LabelSelectorOpIn,
-				Values:   []string{label.PDLabelVal, label.TiKVLabelVal},
-			},
-		},
-	})
-	if err != nil {
-		return err
-	}
-	pvcList, err := oa.kubeCli.CoreV1().PersistentVolumeClaims(ns).List(metav1.ListOptions{LabelSelector: selector.String()})
-	if err != nil {
-		return err
-	}
-	var beforePVCNames []string
-	for _, pvc := range pvcList.Items {
-		beforePVCNames = append(beforePVCNames, pvc.GetName())
-	}
-	log.Logf("%v", beforePVCNames)
-
-	pvList, err := oa.kubeCli.CoreV1().PersistentVolumes().List(metav1.ListOptions{LabelSelector: selector.String()})
-	if err != nil {
-		return err
-	}
-	var beforePVNames []string
-	for _, pv := range pvList.Items {
-		beforePVNames = append(beforePVNames, pv.GetName())
-		log.Logf("%s, %s, %v", pv.Name, pv.Spec.PersistentVolumeReclaimPolicy, pv.Labels)
-		log.Logf("%v", pv.Spec.ClaimRef)
-	}
-	log.Logf("%v", beforePVNames)
-
-	charts := []string{
-		info.ClusterName,
-		fmt.Sprintf("%s-backup", info.ClusterName),
-		fmt.Sprintf("%s-restore", info.ClusterName),
-		fmt.Sprintf("%s-scheduler-backup", info.ClusterName),
-		fmt.Sprintf("%s-%s-drainer", info.ClusterName, DbTypeFile),
-		fmt.Sprintf("%s-%s-drainer", info.ClusterName, DbTypeTiDB),
-		fmt.Sprintf("%s-%s-drainer", info.ClusterName, DbTypeMySQL),
-	}
-	for _, chartName := range charts {
-		res, err := exec.Command("helm", "uninstall", chartName).CombinedOutput()
-		if err != nil && !notFound(string(res)) {
-			return fmt.Errorf("failed to delete chart: %s/%s, %v, %s",
-				info.Namespace, chartName, err, string(res))
-		}
-	}
-
-	time.Sleep(time.Minute)
-
-	pvcList, err = oa.kubeCli.CoreV1().PersistentVolumeClaims(ns).List(metav1.ListOptions{LabelSelector: selector.String()})
-	if err != nil {
-		return err
-	}
-	var afterPVCNames []string
-	for _, pvc := range pvcList.Items {
-		afterPVCNames = append(afterPVCNames, pvc.GetName())
-	}
-	log.Logf("%v", afterPVCNames)
-	if !reflect.DeepEqual(beforePVCNames, afterPVCNames) {
-		return fmt.Errorf("pvc changed when we delete cluster: %s/%s, before: %v, after: %v",
-			ns, tcName, beforePVCNames, afterPVCNames)
-	}
-
-	waitPVFn := func() (done bool, err error) {
-		pvList, err = oa.kubeCli.CoreV1().PersistentVolumes().List(metav1.ListOptions{LabelSelector: selector.String()})
-		if err != nil {
-			return false, nil
-		}
-		var afterPVNames []string
-		for _, pv := range pvList.Items {
-			afterPVNames = append(afterPVNames, pv.GetName())
-		}
-		log.Logf("%v", afterPVNames)
-
-		if !reflect.DeepEqual(beforePVNames, afterPVNames) {
-			log.Logf("ERROR: pv changed when we delete cluster: %s/%s, before: %v, after: %v",
-				ns, tcName, beforePVNames, afterPVNames)
-			return false, nil
-		}
-
-		return true, nil
-	}
-
-	err = wait.Poll(oa.pollInterval, DefaultPollTimeout, waitPVFn)
-	if err != nil {
-		return err
-	}
-
-	err = oa.kubeCli.CoreV1().Pods(info.Namespace).Delete(info.GenerateBackupDirPodName(), &metav1.DeleteOptions{})
-	if err != nil && !errors.IsNotFound(err) {
-		return fmt.Errorf("failed to delete dir pod %v", err)
-	}
-
-	err = oa.kubeCli.CoreV1().Pods(info.Namespace).Delete(blockWriterPodName(info), nil)
-	if err != nil && !errors.IsNotFound(err) {
-		return fmt.Errorf("failed to delete blockwriter pod %v", err)
-	}
-
-	err = oa.kubeCli.CoreV1().Secrets(info.Namespace).Delete(info.InitSecretName, &metav1.DeleteOptions{})
-	if err != nil && !errors.IsNotFound(err) {
-		return fmt.Errorf("failed to delete secret: %s, %v", info.InitSecretName, err)
-	}
-
-	setStr := label.New().Instance(info.ClusterName).String()
-
-	// delete all jobs
-	allJobsSet := label.Label{}.Instance(info.ClusterName).String()
-	if res, err := exec.Command("kubectl", "delete", "jobs", "-n", info.Namespace, "-l", allJobsSet).CombinedOutput(); err != nil {
-		return fmt.Errorf("failed to delete jobs: %v, %s", err, string(res))
-	}
-
-	resources := []string{"pvc"}
-	for _, resource := range resources {
-		if res, err := exec.Command("kubectl", "delete", resource, "-n", info.Namespace, "-l",
-			setStr).CombinedOutput(); err != nil {
-			return fmt.Errorf("failed to delete %s: %v, %s", resource, err, string(res))
-		}
-	}
-
-	// delete pvc of drainer
-	drainerPvcSet := label.Label{}.Instance(info.ClusterName).Component("drainer").String()
-	if res, err := exec.Command("kubectl", "delete", "pvc", "-n", info.Namespace, "-l",
-		drainerPvcSet).CombinedOutput(); err != nil {
-		return fmt.Errorf("failed to delete drainer pvc: %v, %s", err, string(res))
-	}
-
-	// delete all configmaps
-	allConfigMaps := label.New().Instance(info.ClusterName).String()
-	if res, err := exec.Command("kubectl", "delete", "configmaps", "-n", info.Namespace, "-l", allConfigMaps).CombinedOutput(); err != nil {
-		return fmt.Errorf("failed to delete configmaps: %v, %s", err, string(res))
-	}
-
-	err = wait.Poll(10*time.Second, 5*time.Minute, func() (done bool, err error) {
-		patchPVCmd := fmt.Sprintf("kubectl get pv --no-headers -l %s=%s,%s=%s,%s=%s | awk '{print $1}' | "+
-			"xargs -I {} kubectl patch pv {} -p '{\"spec\":{\"persistentVolumeReclaimPolicy\":\"Delete\"}}'",
-			label.ManagedByLabelKey, "tidb-operator",
-			label.NamespaceLabelKey, info.Namespace,
-			label.InstanceLabelKey, info.ClusterName)
-		log.Logf(patchPVCmd)
-		if res, err := exec.Command("/bin/sh", "-c", patchPVCmd).CombinedOutput(); err != nil {
-			log.Logf("failed to patch pv: %v, %s", err, string(res))
-			return false, nil
-		}
-		return true, nil
-	})
-	if err != nil {
-		return err
-	}
-
-	pollFn := func() (bool, error) {
-		if res, err := exec.Command("kubectl", "get", "po", "--output=name", "-n", info.Namespace, "-l", setStr).
-			CombinedOutput(); err != nil || len(res) != 0 {
-			log.Logf("waiting for tidbcluster: %s/%s pods deleting, %v, [%s]",
-				info.Namespace, info.ClusterName, err, string(res))
-			return false, nil
-		}
-
-		pvCmd := fmt.Sprintf("kubectl get pv | grep %s | grep %s 2>/dev/null|grep Released",
-			info.Namespace, info.ClusterName)
-		log.Logf(pvCmd)
-		if res, err := exec.Command("/bin/sh", "-c", pvCmd).CombinedOutput(); len(res) == 0 {
-			return true, nil
-		} else if err != nil {
-			log.Logf("waiting for tidbcluster: %s/%s pv deleting, %v, %s",
-				info.Namespace, info.ClusterName, err, string(res))
-			return false, nil
-		}
-		return true, nil
-	}
-	return wait.PollImmediate(oa.pollInterval, DefaultPollTimeout, pollFn)
-}
-
-// TODO: remove this
-func (oa *operatorActions) CleanTidbClusterOrDie(info *TidbClusterConfig) {
-	if err := oa.CleanTidbCluster(info); err != nil {
-		slack.NotifyAndPanic(err)
-	}
-}
-
-func (oa *operatorActions) CheckTidbClusterStatus(info *TidbClusterConfig) error {
-	log.Logf("checking tidb cluster [%s/%s] status", info.Namespace, info.ClusterName)
-	if info.Clustrer != nil {
-		return oa.crdUtil.WaitForTidbClusterReady(info.Clustrer, 120*time.Minute, 1*time.Minute)
-	}
-
-	ns := info.Namespace
-	tcName := info.ClusterName
-	// TODO: remove redundant checks already in WaitForTidbClusterReady
-	if err := wait.Poll(oa.pollInterval, 10*time.Minute, func() (bool, error) {
-		var tc *v1alpha1.TidbCluster
-		var err error
-		if tc, err = oa.cli.PingcapV1alpha1().TidbClusters(ns).Get(tcName, metav1.GetOptions{}); err != nil {
-			log.Logf("failed to get tidbcluster: %s/%s, %v", ns, tcName, err)
-			return false, nil
-		}
-
-		if b, err := oa.pdMembersReadyFn(tc); !b && err == nil {
-			return false, nil
-		}
-		if b, err := oa.tikvMembersReadyFn(tc); !b && err == nil {
-			return false, nil
-		}
-
-		log.Logf("check tidb cluster begin tidbMembersReadyFn")
-		if b, err := oa.tidbMembersReadyFn(tc); !b && err == nil {
-			return false, nil
-		}
-
-		log.Logf("check tidb cluster begin reclaimPolicySyncFn")
-		if b, err := oa.reclaimPolicySyncFn(tc); !b && err == nil {
-			return false, nil
-		}
-
-		log.Logf("check tidb cluster begin metaSyncFn")
-		if b, err := oa.metaSyncFn(tc); !b && err == nil {
-			return false, nil
-		} else if err != nil {
-			log.Logf("%v", err)
-			return false, nil
-		}
-
-		log.Logf("check tidb cluster begin schedulerHAFn")
-		if b, err := oa.schedulerHAFn(tc); !b && err == nil {
-			return false, nil
-		}
-
-		log.Logf("check all pd and tikv instances have not pod scheduling annotation")
-		if info.OperatorTag != "v1.0.0" {
-			if b, err := oa.podsScheduleAnnHaveDeleted(tc); !b && err == nil {
-				return false, nil
-			}
-		}
-
-		log.Logf("check store labels")
-		if b, err := oa.storeLabelsIsSet(tc, info.TopologyKey); !b && err == nil {
-			return false, nil
-		} else if err != nil {
-			return false, err
-		}
-
-		log.Logf("check tidb cluster begin passwordIsSet")
-		if b, err := oa.passwordIsSet(info); !b && err == nil {
-			return false, nil
-		}
-
-		if info.Monitor {
-			log.Logf("check tidb monitor normal")
-			if b, err := oa.monitorNormal(info); !b && err == nil {
-				return false, nil
-			}
-		}
-		if info.EnableConfigMapRollout {
-			log.Logf("check tidb cluster configuration synced")
-			if b, err := oa.checkTidbClusterConfigUpdated(tc, info); !b && err == nil {
-				return false, nil
-			}
-		}
-		if info.EnablePVReclaim {
-			log.Logf("check reclaim pvs success when scale in pd or tikv")
-			if b, err := oa.checkReclaimPVSuccess(tc); !b && err == nil {
-				return false, nil
-			}
-		}
-		return true, nil
-	}); err != nil {
-		log.Logf("ERROR: check tidb cluster status failed: %s", err.Error())
-		return fmt.Errorf("failed to waiting for tidbcluster %s/%s ready in 120 minutes", ns, tcName)
-	}
-
-	return nil
-}
-
-func (oa *operatorActions) CheckTidbClusterStatusOrDie(info *TidbClusterConfig) {
-	if err := oa.CheckTidbClusterStatus(info); err != nil {
-		slack.NotifyAndPanic(err)
-	}
-}
-
 func (oa *operatorActions) getBlockWriterPod(info *TidbClusterConfig, database string) *corev1.Pod {
 
 	pod := &corev1.Pod{
@@ -1451,913 +1153,6 @@ func (oa *operatorActions) CheckUpgradeOrDie(ctx context.Context, info *TidbClus
 func (oa *operatorActions) DeployMonitor(info *TidbClusterConfig) error { return nil }
 func (oa *operatorActions) CleanMonitor(info *TidbClusterConfig) error  { return nil }
 
-// getMemberContainer gets member container
-func getMemberContainer(kubeCli kubernetes.Interface, stsGetter typedappsv1.StatefulSetsGetter, namespace, tcName, component string) (*corev1.Container, bool) {
-	sts, err := stsGetter.StatefulSets(namespace).Get(fmt.Sprintf("%s-%s", tcName, component), metav1.GetOptions{})
-	if err != nil {
-		log.Logf("ERROR: failed to get sts for component %s of cluster %s/%s", component, namespace, tcName)
-		return nil, false
-	}
-	return getStsContainer(kubeCli, sts, component)
-}
-
-func getStsContainer(kubeCli kubernetes.Interface, sts *apps.StatefulSet, containerName string) (*corev1.Container, bool) {
-	listOption := metav1.ListOptions{
-		LabelSelector: labels.SelectorFromSet(sts.Spec.Selector.MatchLabels).String(),
-	}
-	podList, err := kubeCli.CoreV1().Pods(sts.Namespace).List(listOption)
-	if err != nil {
-		log.Logf("ERROR: fail to get pods for container %s of sts %s/%s", containerName, sts.Namespace, sts.Name)
-		return nil, false
-	}
-	if len(podList.Items) == 0 {
-		log.Logf("ERROR: no pods found for component %s of cluster %s/%s", containerName, sts.Namespace, sts.Name)
-		return nil, false
-	}
-	pod := podList.Items[0]
-	if len(pod.Spec.Containers) == 0 {
-		log.Logf("ERROR: no containers found for component %s of cluster %s/%s", containerName, sts.Namespace, sts.Name)
-		return nil, false
-	}
-	for _, container := range pod.Spec.Containers {
-		if container.Name == containerName {
-			return &container, true
-		}
-	}
-	return nil, false
-}
-
-func (oa *operatorActions) pdMembersReadyFn(tc *v1alpha1.TidbCluster) (bool, error) {
-	if tc.Spec.PD == nil {
-		return true, nil
-	}
-	tcName := tc.GetName()
-	ns := tc.GetNamespace()
-	pdSetName := controller.PDMemberName(tcName)
-
-	pdSet, err := oa.tcStsGetter.StatefulSets(ns).Get(pdSetName, metav1.GetOptions{})
-	if err != nil {
-		log.Logf("failed to get statefulset: %s/%s, %v", ns, pdSetName, err)
-		return false, nil
-	}
-
-	if pdSet.Status.CurrentRevision != pdSet.Status.UpdateRevision {
-		log.Logf("pd sts .Status.CurrentRevision (%s) != .Status.UpdateRevision (%s)", pdSet.Status.CurrentRevision, pdSet.Status.UpdateRevision)
-		return false, nil
-	}
-
-	if !utilstatefulset.IsAllDesiredPodsRunningAndReady(helper.NewHijackClient(oa.kubeCli, oa.asCli), pdSet) {
-		return false, nil
-	}
-
-	if tc.Status.PD.StatefulSet == nil {
-		log.Logf("tidbcluster: %s/%s .status.PD.StatefulSet is nil", ns, tcName)
-		return false, nil
-	}
-	failureCount := len(tc.Status.PD.FailureMembers)
-	replicas := tc.Spec.PD.Replicas + int32(failureCount)
-	if *pdSet.Spec.Replicas != replicas {
-		log.Logf("statefulset: %s/%s .spec.Replicas(%d) != %d",
-			ns, pdSetName, *pdSet.Spec.Replicas, replicas)
-		return false, nil
-	}
-	if pdSet.Status.ReadyReplicas != tc.Spec.PD.Replicas {
-		log.Logf("statefulset: %s/%s .status.ReadyReplicas(%d) != %d",
-			ns, pdSetName, pdSet.Status.ReadyReplicas, tc.Spec.PD.Replicas)
-		return false, nil
-	}
-	if len(tc.Status.PD.Members) != int(tc.Spec.PD.Replicas) {
-		log.Logf("tidbcluster: %s/%s .status.PD.Members count(%d) != %d",
-			ns, tcName, len(tc.Status.PD.Members), tc.Spec.PD.Replicas)
-		return false, nil
-	}
-	if pdSet.Status.ReadyReplicas != pdSet.Status.Replicas {
-		log.Logf("statefulset: %s/%s .status.ReadyReplicas(%d) != .status.Replicas(%d)",
-			ns, pdSetName, pdSet.Status.ReadyReplicas, pdSet.Status.Replicas)
-		return false, nil
-	}
-
-	c, found := getMemberContainer(oa.kubeCli, oa.tcStsGetter, ns, tc.Name, label.PDLabelVal)
-	if !found {
-		log.Logf("statefulset: %s/%s not found containers[name=pd] or pod %s-0",
-			ns, pdSetName, pdSetName)
-		return false, nil
-	}
-
-	if tc.PDImage() != c.Image {
-		log.Logf("statefulset: %s/%s .spec.template.spec.containers[name=pd].image(%s) != %s",
-			ns, pdSetName, c.Image, tc.PDImage())
-		return false, nil
-	}
-
-	for _, member := range tc.Status.PD.Members {
-		if !member.Health {
-			log.Logf("tidbcluster: %s/%s pd member(%s/%s) is not health",
-				ns, tcName, member.ID, member.Name)
-			return false, nil
-		}
-	}
-
-	pdServiceName := controller.PDMemberName(tcName)
-	pdPeerServiceName := controller.PDPeerMemberName(tcName)
-	if _, err := oa.kubeCli.CoreV1().Services(ns).Get(pdServiceName, metav1.GetOptions{}); err != nil {
-		log.Logf("failed to get service: %s/%s", ns, pdServiceName)
-		return false, nil
-	}
-	if _, err := oa.kubeCli.CoreV1().Services(ns).Get(pdPeerServiceName, metav1.GetOptions{}); err != nil {
-		log.Logf("failed to get peer service: %s/%s", ns, pdPeerServiceName)
-		return false, nil
-	}
-
-	return true, nil
-}
-
-func (oa *operatorActions) tikvMembersReadyFn(tc *v1alpha1.TidbCluster) (bool, error) {
-	if tc.Spec.TiKV == nil {
-		return true, nil
-	}
-	tcName := tc.GetName()
-	ns := tc.GetNamespace()
-	tikvSetName := controller.TiKVMemberName(tcName)
-
-	tikvSet, err := oa.tcStsGetter.StatefulSets(ns).Get(tikvSetName, metav1.GetOptions{})
-	if err != nil {
-		log.Logf("failed to get statefulset: %s/%s, %v", ns, tikvSetName, err)
-		return false, nil
-	}
-
-	if tikvSet.Status.CurrentRevision != tikvSet.Status.UpdateRevision {
-		log.Logf("tikv sts .Status.CurrentRevision (%s) != .Status.UpdateRevision (%s)", tikvSet.Status.CurrentRevision, tikvSet.Status.UpdateRevision)
-		return false, nil
-	}
-
-	if !utilstatefulset.IsAllDesiredPodsRunningAndReady(helper.NewHijackClient(oa.kubeCli, oa.asCli), tikvSet) {
-		return false, nil
-	}
-
-	if tc.Status.TiKV.StatefulSet == nil {
-		log.Logf("tidbcluster: %s/%s .status.TiKV.StatefulSet is nil", ns, tcName)
-		return false, nil
-	}
-	failureCount := len(tc.Status.TiKV.FailureStores)
-	replicas := tc.Spec.TiKV.Replicas + int32(failureCount)
-	if *tikvSet.Spec.Replicas != replicas {
-		log.Logf("statefulset: %s/%s .spec.Replicas(%d) != %d",
-			ns, tikvSetName, *tikvSet.Spec.Replicas, replicas)
-		return false, nil
-	}
-	if tikvSet.Status.ReadyReplicas != replicas {
-		log.Logf("statefulset: %s/%s .status.ReadyReplicas(%d) != %d",
-			ns, tikvSetName, tikvSet.Status.ReadyReplicas, replicas)
-		return false, nil
-	}
-	if len(tc.Status.TiKV.Stores) != int(replicas) {
-		log.Logf("tidbcluster: %s/%s .status.TiKV.Stores.count(%d) != %d",
-			ns, tcName, len(tc.Status.TiKV.Stores), replicas)
-		return false, nil
-	}
-	if tikvSet.Status.ReadyReplicas != tikvSet.Status.Replicas {
-		log.Logf("statefulset: %s/%s .status.ReadyReplicas(%d) != .status.Replicas(%d)",
-			ns, tikvSetName, tikvSet.Status.ReadyReplicas, tikvSet.Status.Replicas)
-		return false, nil
-	}
-
-	c, found := getMemberContainer(oa.kubeCli, oa.tcStsGetter, ns, tc.Name, label.TiKVLabelVal)
-	if !found {
-		log.Logf("statefulset: %s/%s not found containers[name=tikv] or pod %s-0",
-			ns, tikvSetName, tikvSetName)
-		return false, nil
-	}
-
-	if tc.TiKVImage() != c.Image {
-		log.Logf("statefulset: %s/%s .spec.template.spec.containers[name=tikv].image(%s) != %s",
-			ns, tikvSetName, c.Image, tc.TiKVImage())
-		return false, nil
-	}
-
-	for _, store := range tc.Status.TiKV.Stores {
-		if store.State != v1alpha1.TiKVStateUp {
-			log.Logf("tidbcluster: %s/%s's store(%s) state != %s", ns, tcName, store.ID, v1alpha1.TiKVStateUp)
-			return false, nil
-		}
-	}
-
-	tikvPeerServiceName := controller.TiKVPeerMemberName(tcName)
-	if _, err := oa.kubeCli.CoreV1().Services(ns).Get(tikvPeerServiceName, metav1.GetOptions{}); err != nil {
-		log.Logf("failed to get peer service: %s/%s", ns, tikvPeerServiceName)
-		return false, nil
-	}
-
-	return true, nil
-}
-
-func (oa *operatorActions) tiflashMembersReadyFn(tc *v1alpha1.TidbCluster) (bool, error) {
-	tcName := tc.GetName()
-	ns := tc.GetNamespace()
-	tiflashSetName := controller.TiFlashMemberName(tcName)
-
-	tiflashSet, err := oa.tcStsGetter.StatefulSets(ns).Get(tiflashSetName, metav1.GetOptions{})
-	if err != nil {
-		log.Logf("TiFlash failed to get statefulset: %s/%s, %v", ns, tiflashSetName, err)
-		return false, nil
-	}
-
-	if tiflashSet.Status.CurrentRevision != tiflashSet.Status.UpdateRevision {
-		log.Logf("tiflash sts .Status.CurrentRevision (%s) != .Status.UpdateRevision (%s)", tiflashSet.Status.CurrentRevision, tiflashSet.Status.UpdateRevision)
-		return false, nil
-	}
-
-	if !utilstatefulset.IsAllDesiredPodsRunningAndReady(helper.NewHijackClient(oa.kubeCli, oa.asCli), tiflashSet) {
-		return false, nil
-	}
-
-	if tc.Status.TiFlash.StatefulSet == nil {
-		log.Logf("tidbcluster: %s/%s .status.TiFlash.StatefulSet is nil", ns, tcName)
-		return false, nil
-	}
-	failureCount := len(tc.Status.TiFlash.FailureStores)
-	replicas := tc.Spec.TiFlash.Replicas + int32(failureCount)
-	if *tiflashSet.Spec.Replicas != replicas {
-		log.Logf("TiFlash statefulset: %s/%s .spec.Replicas(%d) != %d",
-			ns, tiflashSetName, *tiflashSet.Spec.Replicas, replicas)
-		return false, nil
-	}
-	if tiflashSet.Status.ReadyReplicas != replicas {
-		log.Logf("TiFlash statefulset: %s/%s .status.ReadyReplicas(%d) != %d",
-			ns, tiflashSetName, tiflashSet.Status.ReadyReplicas, replicas)
-		return false, nil
-	}
-	if len(tc.Status.TiFlash.Stores) != int(replicas) {
-		log.Logf("tidbcluster: %s/%s .status.TiFlash.Stores.count(%d) != %d",
-			ns, tcName, len(tc.Status.TiFlash.Stores), replicas)
-		return false, nil
-	}
-	if tiflashSet.Status.ReadyReplicas != tiflashSet.Status.Replicas {
-		log.Logf("statefulset: %s/%s .status.ReadyReplicas(%d) != .status.Replicas(%d)",
-			ns, tiflashSetName, tiflashSet.Status.ReadyReplicas, tiflashSet.Status.Replicas)
-		return false, nil
-	}
-
-	c, found := getMemberContainer(oa.kubeCli, oa.tcStsGetter, ns, tc.Name, label.TiFlashLabelVal)
-	if !found {
-		log.Logf("statefulset: %s/%s not found containers[name=tiflash] or pod %s-0",
-			ns, tiflashSetName, tiflashSetName)
-		return false, nil
-	}
-
-	if tc.TiFlashImage() != c.Image {
-		log.Logf("statefulset: %s/%s .spec.template.spec.containers[name=tiflash].image(%s) != %s",
-			ns, tiflashSetName, c.Image, tc.TiFlashImage())
-		return false, nil
-	}
-
-	for _, store := range tc.Status.TiFlash.Stores {
-		if store.State != v1alpha1.TiKVStateUp {
-			log.Logf("TiFlash tidbcluster: %s/%s's store(%s) state != %s", ns, tcName, store.ID, v1alpha1.TiKVStateUp)
-			return false, nil
-		}
-	}
-
-	tiflashPeerServiceName := controller.TiFlashPeerMemberName(tcName)
-	if _, err := oa.kubeCli.CoreV1().Services(ns).Get(tiflashPeerServiceName, metav1.GetOptions{}); err != nil {
-		log.Logf("failed to get peer service: %s/%s", ns, tiflashPeerServiceName)
-		return false, nil
-	}
-	log.Logf("TiFlash ready: %s/%s", ns, tcName)
-	return true, nil
-}
-
-func (oa *operatorActions) tidbMembersReadyFn(tc *v1alpha1.TidbCluster) (bool, error) {
-	if tc.Spec.TiDB == nil {
-		return true, nil
-	}
-	tcName := tc.GetName()
-	ns := tc.GetNamespace()
-	tidbSetName := controller.TiDBMemberName(tcName)
-
-	tidbSet, err := oa.tcStsGetter.StatefulSets(ns).Get(tidbSetName, metav1.GetOptions{})
-	if err != nil {
-		log.Logf("failed to get statefulset: %s/%s, %v", ns, tidbSetName, err)
-		return false, nil
-	}
-
-	if tidbSet.Status.CurrentRevision != tidbSet.Status.UpdateRevision {
-		log.Logf("tidb sts .Status.CurrentRevision (%s) != .Status.UpdateRevision (%s)", tidbSet.Status.CurrentRevision, tidbSet.Status.UpdateRevision)
-		return false, nil
-	}
-
-	if !utilstatefulset.IsAllDesiredPodsRunningAndReady(helper.NewHijackClient(oa.kubeCli, oa.asCli), tidbSet) {
-		return false, nil
-	}
-
-	if tc.Status.TiDB.StatefulSet == nil {
-		log.Logf("tidbcluster: %s/%s .status.TiDB.StatefulSet is nil", ns, tcName)
-		return false, nil
-	}
-	failureCount := len(tc.Status.TiDB.FailureMembers)
-	replicas := tc.Spec.TiDB.Replicas + int32(failureCount)
-	if *tidbSet.Spec.Replicas != replicas {
-		log.Logf("statefulset: %s/%s .spec.Replicas(%d) != %d",
-			ns, tidbSetName, *tidbSet.Spec.Replicas, replicas)
-		return false, nil
-	}
-	if tidbSet.Status.ReadyReplicas != tc.Spec.TiDB.Replicas {
-		log.Logf("statefulset: %s/%s .status.ReadyReplicas(%d) != %d",
-			ns, tidbSetName, tidbSet.Status.ReadyReplicas, tc.Spec.TiDB.Replicas)
-		return false, nil
-	}
-	if len(tc.Status.TiDB.Members) != int(tc.Spec.TiDB.Replicas) {
-		log.Logf("tidbcluster: %s/%s .status.TiDB.Members count(%d) != %d",
-			ns, tcName, len(tc.Status.TiDB.Members), tc.Spec.TiDB.Replicas)
-		return false, nil
-	}
-	if tidbSet.Status.ReadyReplicas != tidbSet.Status.Replicas {
-		log.Logf("statefulset: %s/%s .status.ReadyReplicas(%d) != .status.Replicas(%d)",
-			ns, tidbSetName, tidbSet.Status.ReadyReplicas, tidbSet.Status.Replicas)
-		return false, nil
-	}
-
-	c, found := getMemberContainer(oa.kubeCli, oa.tcStsGetter, ns, tc.Name, label.TiDBLabelVal)
-	if !found {
-		log.Logf("statefulset: %s/%s not found containers[name=tidb] or pod %s-0",
-			ns, tidbSetName, tidbSetName)
-		return false, nil
-	}
-
-	if tc.TiDBImage() != c.Image {
-		log.Logf("statefulset: %s/%s .spec.template.spec.containers[name=tidb].image(%s) != %s",
-			ns, tidbSetName, c.Image, tc.TiDBImage())
-		return false, nil
-	}
-
-	_, err = oa.kubeCli.CoreV1().Services(ns).Get(tidbSetName, metav1.GetOptions{})
-	if err != nil {
-		log.Logf("failed to get service: %s/%s", ns, tidbSetName)
-		return false, nil
-	}
-	_, err = oa.kubeCli.CoreV1().Services(ns).Get(controller.TiDBPeerMemberName(tcName), metav1.GetOptions{})
-	if err != nil {
-		log.Logf("failed to get peer service: %s/%s", ns, controller.TiDBPeerMemberName(tcName))
-		return false, nil
-	}
-
-	return true, nil
-}
-
-func (oa *operatorActions) reclaimPolicySyncFn(tc *v1alpha1.TidbCluster) (bool, error) {
-	ns := tc.GetNamespace()
-	tcName := tc.GetName()
-	listOptions := metav1.ListOptions{
-		LabelSelector: labels.SelectorFromSet(
-			label.New().Instance(tcName).Labels(),
-		).String(),
-	}
-	var pvcList *corev1.PersistentVolumeClaimList
-	var err error
-	if pvcList, err = oa.kubeCli.CoreV1().PersistentVolumeClaims(ns).List(listOptions); err != nil {
-		log.Logf("failed to list pvs for tidbcluster %s/%s, %v", ns, tcName, err)
-		return false, nil
-	}
-
-	for _, pvc := range pvcList.Items {
-		pvName := pvc.Spec.VolumeName
-		if pv, err := oa.kubeCli.CoreV1().PersistentVolumes().Get(pvName, metav1.GetOptions{}); err != nil {
-			log.Logf("failed to get pv: %s, error: %v", pvName, err)
-			return false, nil
-		} else if pv.Spec.PersistentVolumeReclaimPolicy != *tc.Spec.PVReclaimPolicy {
-			log.Logf("pv: %s's reclaimPolicy is not Retain", pvName)
-			return false, nil
-		}
-	}
-
-	return true, nil
-}
-
-func (oa *operatorActions) metaSyncFn(tc *v1alpha1.TidbCluster) (bool, error) {
-	ns := tc.GetNamespace()
-	tcName := tc.GetName()
-
-	pdClient, cancel, err := oa.getPDClient(tc)
-	if err != nil {
-		log.Logf("failed to create external PD client for tidb cluster %q: %v", tc.GetName(), err)
-		return false, nil
-	}
-	defer cancel()
-	var cluster *metapb.Cluster
-	if cluster, err = pdClient.GetCluster(); err != nil {
-		log.Logf("failed to get cluster from pdControl: %s/%s, error: %v", ns, tcName, err)
-		return false, nil
-	}
-
-	clusterID := strconv.FormatUint(cluster.Id, 10)
-	listOptions := metav1.ListOptions{
-		LabelSelector: labels.SelectorFromSet(
-			label.New().Instance(tcName).Labels(),
-		).String(),
-	}
-
-	var podList *corev1.PodList
-	if podList, err = oa.kubeCli.CoreV1().Pods(ns).List(listOptions); err != nil {
-		log.Logf("failed to list pods for tidbcluster %s/%s, %v", ns, tcName, err)
-		return false, nil
-	}
-
-outerLoop:
-	for _, pod := range podList.Items {
-		podName := pod.GetName()
-		if pod.Labels[label.ClusterIDLabelKey] != clusterID {
-			log.Logf("tidbcluster %s/%s's pod %s's label %s not equals %s ",
-				ns, tcName, podName, label.ClusterIDLabelKey, clusterID)
-			return false, nil
-		}
-
-		component := pod.Labels[label.ComponentLabelKey]
-		switch component {
-		case label.PDLabelVal:
-			var memberID string
-			members, err := pdClient.GetMembers()
-			if err != nil {
-				log.Logf("failed to get members for tidbcluster %s/%s, %v", ns, tcName, err)
-				return false, nil
-			}
-			for _, member := range members.Members {
-				if member.Name == podName {
-					memberID = strconv.FormatUint(member.GetMemberId(), 10)
-					break
-				}
-			}
-			if memberID == "" {
-				log.Logf("tidbcluster: %s/%s's pod %s label [%s] is empty",
-					ns, tcName, podName, label.MemberIDLabelKey)
-				return false, nil
-			}
-			if pod.Labels[label.MemberIDLabelKey] != memberID {
-				return false, fmt.Errorf("tidbcluster: %s/%s's pod %s label [%s] not equals %s",
-					ns, tcName, podName, label.MemberIDLabelKey, memberID)
-			}
-		case label.TiKVLabelVal:
-			var storeID string
-			stores, err := pdClient.GetStores()
-			if err != nil {
-				log.Logf("failed to get stores for tidbcluster %s/%s, %v", ns, tcName, err)
-				return false, nil
-			}
-			for _, store := range stores.Stores {
-				addr := store.Store.GetAddress()
-				if strings.Split(addr, ".")[0] == podName {
-					storeID = strconv.FormatUint(store.Store.GetId(), 10)
-					break
-				}
-			}
-			if storeID == "" {
-				log.Logf("tidbcluster: %s/%s's pod %s label [%s] is empty",
-					tc.GetNamespace(), tc.GetName(), podName, label.StoreIDLabelKey)
-				return false, nil
-			}
-			if pod.Labels[label.StoreIDLabelKey] != storeID {
-				return false, fmt.Errorf("tidbcluster: %s/%s's pod %s label [%s] not equals %s",
-					ns, tcName, podName, label.StoreIDLabelKey, storeID)
-			}
-		case label.TiDBLabelVal:
-			continue outerLoop
-		default:
-			continue outerLoop
-		}
-
-		var pvcName string
-		for _, vol := range pod.Spec.Volumes {
-			if vol.PersistentVolumeClaim != nil {
-				pvcName = vol.PersistentVolumeClaim.ClaimName
-				break
-			}
-		}
-		if pvcName == "" {
-			return false, fmt.Errorf("pod: %s/%s's pvcName is empty", ns, podName)
-		}
-
-		var pvc *corev1.PersistentVolumeClaim
-		if pvc, err = oa.kubeCli.CoreV1().PersistentVolumeClaims(ns).Get(pvcName, metav1.GetOptions{}); err != nil {
-			log.Logf("failed to get pvc %s/%s for pod %s/%s", ns, pvcName, ns, podName)
-			return false, nil
-		}
-		if pvc.Labels[label.ClusterIDLabelKey] != clusterID {
-			return false, fmt.Errorf("tidbcluster: %s/%s's pvc %s label [%s] not equals %s ",
-				ns, tcName, pvcName, label.ClusterIDLabelKey, clusterID)
-		}
-		if pvc.Labels[label.MemberIDLabelKey] != pod.Labels[label.MemberIDLabelKey] {
-			return false, fmt.Errorf("tidbcluster: %s/%s's pvc %s label [%s=%s] not equals pod lablel [%s=%s]",
-				ns, tcName, pvcName,
-				label.MemberIDLabelKey, pvc.Labels[label.MemberIDLabelKey],
-				label.MemberIDLabelKey, pod.Labels[label.MemberIDLabelKey])
-		}
-		if pvc.Labels[label.StoreIDLabelKey] != pod.Labels[label.StoreIDLabelKey] {
-			return false, fmt.Errorf("tidbcluster: %s/%s's pvc %s label[%s=%s] not equals pod lable[%s=%s]",
-				ns, tcName, pvcName,
-				label.StoreIDLabelKey, pvc.Labels[label.StoreIDLabelKey],
-				label.StoreIDLabelKey, pod.Labels[label.StoreIDLabelKey])
-		}
-		if pvc.Annotations[label.AnnPodNameKey] != podName {
-			return false, fmt.Errorf("tidbcluster: %s/%s's pvc %s annotations [%s] not equals podName: %s",
-				ns, tcName, pvcName, label.AnnPodNameKey, podName)
-		}
-
-		pvName := pvc.Spec.VolumeName
-		var pv *corev1.PersistentVolume
-		if pv, err = oa.kubeCli.CoreV1().PersistentVolumes().Get(pvName, metav1.GetOptions{}); err != nil {
-			log.Logf("failed to get pv for pvc %s/%s, %v", ns, pvcName, err)
-			return false, nil
-		}
-		if pv.Labels[label.NamespaceLabelKey] != ns {
-			return false, fmt.Errorf("tidbcluster: %s/%s 's pv %s label [%s] not equals %s",
-				ns, tcName, pvName, label.NamespaceLabelKey, ns)
-		}
-		if pv.Labels[label.ComponentLabelKey] != pod.Labels[label.ComponentLabelKey] {
-			return false, fmt.Errorf("tidbcluster: %s/%s's pv %s label [%s=%s] not equals pod label[%s=%s]",
-				ns, tcName, pvName,
-				label.ComponentLabelKey, pv.Labels[label.ComponentLabelKey],
-				label.ComponentLabelKey, pod.Labels[label.ComponentLabelKey])
-		}
-		if pv.Labels[label.NameLabelKey] != pod.Labels[label.NameLabelKey] {
-			return false, fmt.Errorf("tidbcluster: %s/%s's pv %s label [%s=%s] not equals pod label [%s=%s]",
-				ns, tcName, pvName,
-				label.NameLabelKey, pv.Labels[label.NameLabelKey],
-				label.NameLabelKey, pod.Labels[label.NameLabelKey])
-		}
-		if pv.Labels[label.ManagedByLabelKey] != pod.Labels[label.ManagedByLabelKey] {
-			return false, fmt.Errorf("tidbcluster: %s/%s's pv %s label [%s=%s] not equals pod label [%s=%s]",
-				ns, tcName, pvName,
-				label.ManagedByLabelKey, pv.Labels[label.ManagedByLabelKey],
-				label.ManagedByLabelKey, pod.Labels[label.ManagedByLabelKey])
-		}
-		if pv.Labels[label.InstanceLabelKey] != pod.Labels[label.InstanceLabelKey] {
-			return false, fmt.Errorf("tidbcluster: %s/%s's pv %s label [%s=%s] not equals pod label [%s=%s]",
-				ns, tcName, pvName,
-				label.InstanceLabelKey, pv.Labels[label.InstanceLabelKey],
-				label.InstanceLabelKey, pod.Labels[label.InstanceLabelKey])
-		}
-		if pv.Labels[label.ClusterIDLabelKey] != clusterID {
-			return false, fmt.Errorf("tidbcluster: %s/%s's pv %s label [%s] not equals %s",
-				ns, tcName, pvName, label.ClusterIDLabelKey, clusterID)
-		}
-		if pv.Labels[label.MemberIDLabelKey] != pod.Labels[label.MemberIDLabelKey] {
-			return false, fmt.Errorf("tidbcluster: %s/%s's pv %s label [%s=%s] not equals pod label [%s=%s]",
-				ns, tcName, pvName,
-				label.MemberIDLabelKey, pv.Labels[label.MemberIDLabelKey],
-				label.MemberIDLabelKey, pod.Labels[label.MemberIDLabelKey])
-		}
-		if pv.Labels[label.StoreIDLabelKey] != pod.Labels[label.StoreIDLabelKey] {
-			return false, fmt.Errorf("tidbcluster: %s/%s's pv %s label [%s=%s] not equals pod label [%s=%s]",
-				ns, tcName, pvName,
-				label.StoreIDLabelKey, pv.Labels[label.StoreIDLabelKey],
-				label.StoreIDLabelKey, pod.Labels[label.StoreIDLabelKey])
-		}
-		if pv.Annotations[label.AnnPodNameKey] != podName {
-			return false, fmt.Errorf("tidbcluster:[%s/%s's pv %s annotations [%s] not equals %s",
-				ns, tcName, pvName, label.AnnPodNameKey, podName)
-		}
-	}
-
-	return true, nil
-}
-
-func (oa *operatorActions) schedulerHAFn(tc *v1alpha1.TidbCluster) (bool, error) {
-	ns := tc.GetNamespace()
-	tcName := tc.GetName()
-
-	fn := func(component string) (bool, error) {
-		nodeMap := make(map[string][]string)
-		listOptions := metav1.ListOptions{
-			LabelSelector: labels.SelectorFromSet(
-				label.New().Instance(tcName).Component(component).Labels()).String(),
-		}
-		var podList *corev1.PodList
-		var err error
-		if podList, err = oa.kubeCli.CoreV1().Pods(ns).List(listOptions); err != nil {
-			log.Logf("failed to list pods for tidbcluster %s/%s, %v", ns, tcName, err)
-			return false, nil
-		}
-
-		totalCount := len(podList.Items)
-		for _, pod := range podList.Items {
-			nodeName := pod.Spec.NodeName
-			if len(nodeMap[nodeName]) == 0 {
-				nodeMap[nodeName] = make([]string, 0)
-			}
-			nodeMap[nodeName] = append(nodeMap[nodeName], pod.GetName())
-			if len(nodeMap[nodeName]) > totalCount/2 {
-				return false, fmt.Errorf("node %s have %d pods, greater than %d/2",
-					nodeName, len(nodeMap[nodeName]), totalCount)
-			}
-		}
-		return true, nil
-	}
-
-	components := []string{label.PDLabelVal, label.TiKVLabelVal}
-	for _, com := range components {
-		if b, err := fn(com); err != nil {
-			return false, err
-		} else if !b && err == nil {
-			return false, nil
-		}
-	}
-
-	return true, nil
-}
-
-func (oa *operatorActions) podsScheduleAnnHaveDeleted(tc *v1alpha1.TidbCluster) (bool, error) {
-	ns := tc.GetNamespace()
-	tcName := tc.GetName()
-
-	listOptions := metav1.ListOptions{
-		LabelSelector: labels.SelectorFromSet(
-			label.New().Instance(tcName).Labels()).String(),
-	}
-
-	pvcList, err := oa.kubeCli.CoreV1().PersistentVolumeClaims(ns).List(listOptions)
-	if err != nil {
-		log.Logf("failed to list pvcs for tidb cluster %s/%s, err: %v", ns, tcName, err)
-		return false, nil
-	}
-
-	for _, pvc := range pvcList.Items {
-		pvcName := pvc.GetName()
-		l := label.Label(pvc.Labels)
-		if !(l.IsPD() || l.IsTiKV()) {
-			continue
-		}
-
-		if _, exist := pvc.Annotations[label.AnnPVCPodScheduling]; exist {
-			log.Logf("tidb cluster %s/%s pvc %s has pod scheduling annotation", ns, tcName, pvcName)
-			return false, nil
-		}
-	}
-
-	return true, nil
-}
-
-func (oa *operatorActions) checkReclaimPVSuccess(tc *v1alpha1.TidbCluster) (bool, error) {
-	// check pv reclaim	for pd
-	if err := oa.checkComponentReclaimPVSuccess(tc, label.PDLabelVal); err != nil {
-		log.Logf("%v", err)
-		return false, nil
-	}
-
-	// check pv reclaim for tikv
-	if err := oa.checkComponentReclaimPVSuccess(tc, label.TiKVLabelVal); err != nil {
-		log.Logf("%v", err)
-		return false, nil
-	}
-	return true, nil
-}
-
-func (oa *operatorActions) checkComponentReclaimPVSuccess(tc *v1alpha1.TidbCluster, component string) error {
-	ns := tc.GetNamespace()
-	tcName := tc.GetName()
-
-	var replica int
-	switch component {
-	case label.PDLabelVal:
-		replica = int(tc.Spec.PD.Replicas)
-	case label.TiKVLabelVal:
-		replica = int(tc.Spec.TiKV.Replicas)
-	default:
-		return fmt.Errorf("check tidb cluster %s/%s component %s is not supported", ns, tcName, component)
-	}
-
-	pvcList, err := oa.getComponentPVCList(tc, component)
-	if err != nil {
-		return err
-	}
-
-	pvList, err := oa.getComponentPVList(tc, component)
-	if err != nil {
-		return err
-	}
-
-	if len(pvcList) != replica {
-		return fmt.Errorf("tidb cluster %s/%s component %s pvc has not been reclaimed completely, expected: %d, got %d", ns, tcName, component, replica, len(pvcList))
-	}
-
-	if len(pvList) != replica {
-		return fmt.Errorf("tidb cluster %s/%s component %s pv has not been reclaimed completely, expected: %d, got %d", ns, tcName, component, replica, len(pvList))
-	}
-
-	return nil
-}
-
-func (oa *operatorActions) getComponentPVCList(tc *v1alpha1.TidbCluster, component string) ([]corev1.PersistentVolumeClaim, error) {
-	ns := tc.GetNamespace()
-	tcName := tc.GetName()
-
-	listOptions := metav1.ListOptions{
-		LabelSelector: labels.SelectorFromSet(
-			label.New().Instance(tcName).Component(component).Labels()).String(),
-	}
-
-	pvcList, err := oa.kubeCli.CoreV1().PersistentVolumeClaims(ns).List(listOptions)
-	if err != nil {
-		err := fmt.Errorf("failed to list pvcs for tidb cluster %s/%s, component: %s, err: %v", ns, tcName, component, err)
-		return nil, err
-	}
-	return pvcList.Items, nil
-}
-
-func (oa *operatorActions) getComponentPVList(tc *v1alpha1.TidbCluster, component string) ([]corev1.PersistentVolume, error) {
-	ns := tc.GetNamespace()
-	tcName := tc.GetName()
-
-	listOptions := metav1.ListOptions{
-		LabelSelector: labels.SelectorFromSet(
-			label.New().Instance(tcName).Component(component).Namespace(ns).Labels()).String(),
-	}
-
-	pvList, err := oa.kubeCli.CoreV1().PersistentVolumes().List(listOptions)
-	if err != nil {
-		err := fmt.Errorf("failed to list pvs for tidb cluster %s/%s, component: %s, err: %v", ns, tcName, component, err)
-		return nil, err
-	}
-	return pvList.Items, nil
-}
-
-func (oa *operatorActions) storeLabelsIsSet(tc *v1alpha1.TidbCluster, topologyKey string) (bool, error) {
-	pdClient, cancel, err := oa.getPDClient(tc)
-	if err != nil {
-		log.Logf("failed to create external PD client for tidb cluster %q: %v", tc.GetName(), err)
-		return false, nil
-	}
-	defer cancel()
-	for _, store := range tc.Status.TiKV.Stores {
-		storeID, err := strconv.ParseUint(store.ID, 10, 64)
-		if err != nil {
-			return false, err
-		}
-		storeInfo, err := pdClient.GetStore(storeID)
-		if err != nil {
-			return false, nil
-		}
-		if len(storeInfo.Store.Labels) == 0 {
-			return false, nil
-		}
-		for _, label := range storeInfo.Store.Labels {
-			if label.Key != topologyKey {
-				return false, nil
-			}
-		}
-	}
-	return true, nil
-}
-
-func (oa *operatorActions) passwordIsSet(clusterInfo *TidbClusterConfig) (bool, error) {
-	ns := clusterInfo.Namespace
-	tcName := clusterInfo.ClusterName
-	jobName := tcName + "-tidb-initializer"
-
-	var job *batchv1.Job
-	var err error
-	if job, err = oa.kubeCli.BatchV1().Jobs(ns).Get(jobName, metav1.GetOptions{}); err != nil {
-		log.Logf("failed to get job %s/%s, %v", ns, jobName, err)
-		return false, nil
-	}
-	if job.Status.Succeeded < 1 {
-		log.Logf("tidbcluster: %s/%s password setter job not finished", ns, tcName)
-		return false, nil
-	}
-
-	var db *sql.DB
-	dsn, cancel, err := oa.getTiDBDSN(ns, tcName, "test", clusterInfo.Password)
-	if err != nil {
-		log.Logf("failed to get TiDB DSN: %v", err)
-		return false, nil
-	}
-	defer cancel()
-	if db, err = sql.Open("mysql", dsn); err != nil {
-		log.Logf("can't open connection to mysql: %s, %v", dsn, err)
-		return false, nil
-	}
-	defer db.Close()
-	if err := db.Ping(); err != nil {
-		log.Logf("can't connect to mysql: %s with password %s, %v", dsn, clusterInfo.Password, err)
-		return false, nil
-	}
-
-	return true, nil
-}
-
-func (oa *operatorActions) monitorNormal(clusterInfo *TidbClusterConfig) (bool, error) {
-	ns := clusterInfo.Namespace
-	tcName := clusterInfo.ClusterName
-	monitorDeploymentName := fmt.Sprintf("%s-monitor", tcName)
-	monitorDeployment, err := oa.kubeCli.AppsV1().Deployments(ns).Get(monitorDeploymentName, metav1.GetOptions{})
-	if err != nil {
-		log.Logf("get monitor deployment: [%s/%s] failed", ns, monitorDeploymentName)
-		return false, nil
-	}
-	if monitorDeployment.Status.ReadyReplicas < 1 {
-		log.Logf("monitor ready replicas %d < 1", monitorDeployment.Status.ReadyReplicas)
-		return false, nil
-	}
-	if err := oa.checkPrometheus(clusterInfo); err != nil {
-		log.Logf("check [%s/%s]'s prometheus data failed: %v", ns, monitorDeploymentName, err)
-		return false, nil
-	}
-
-	if err := oa.checkGrafanaData(clusterInfo); err != nil {
-		log.Logf("check [%s/%s]'s grafana data failed: %v", ns, monitorDeploymentName, err)
-		return false, nil
-	}
-	return true, nil
-}
-
-func (oa *operatorActions) checkTidbClusterConfigUpdated(tc *v1alpha1.TidbCluster, clusterInfo *TidbClusterConfig) (bool, error) {
-	if ok := oa.checkPdConfigUpdated(tc, clusterInfo); !ok {
-		return false, nil
-	}
-	if ok := oa.checkTiKVConfigUpdated(tc, clusterInfo); !ok {
-		return false, nil
-	}
-	if ok := oa.checkTiDBConfigUpdated(tc, clusterInfo); !ok {
-		return false, nil
-	}
-	return true, nil
-}
-
-func (oa *operatorActions) checkPdConfigUpdated(tc *v1alpha1.TidbCluster, clusterInfo *TidbClusterConfig) bool {
-	pdClient, cancel, err := oa.getPDClient(tc)
-	if err != nil {
-		log.Logf("failed to create external PD client for tidb cluster %q: %v", tc.GetName(), err)
-		return false
-	}
-	defer cancel()
-	config, err := pdClient.GetConfig()
-	if err != nil {
-		log.Logf("failed to get PD configuraion from tidb cluster [%s/%s]", tc.Namespace, tc.Name)
-		return false
-	}
-	if len(clusterInfo.PDLogLevel) > 0 && clusterInfo.PDLogLevel != config.Log.Level {
-		log.Logf("check [%s/%s] PD logLevel configuration updated failed: desired [%s], actual [%s] not equal",
-			tc.Namespace,
-			tc.Name,
-			clusterInfo.PDLogLevel,
-			config.Log.Level)
-		return false
-	}
-	// TODO: fix #487 PD configuration update for persisted configurations
-	//if clusterInfo.PDMaxReplicas > 0 && config.Replication.MaxReplicas != uint64(clusterInfo.PDMaxReplicas) {
-	//	log.Logf("check [%s/%s] PD maxReplicas configuration updated failed: desired [%d], actual [%d] not equal",
-	//		tc.Namespace,
-	//		tc.Name,
-	//		clusterInfo.PDMaxReplicas,
-	//		config.Replication.MaxReplicas)
-	//	return false
-	//}
-	return true
-}
-
-func (oa *operatorActions) checkTiDBConfigUpdated(tc *v1alpha1.TidbCluster, clusterInfo *TidbClusterConfig) bool {
-	ordinals, err := util.GetPodOrdinals(tc, v1alpha1.TiDBMemberType)
-	if err != nil {
-		log.Logf("failed to get pod ordinals for tidb cluster %s/%s (member: %v)", tc.Namespace, tc.Name, v1alpha1.TiDBMemberType)
-		return false
-	}
-	for i := range ordinals {
-		config, err := oa.tidbControl.GetSettings(tc, int32(i))
-		if err != nil {
-			log.Logf("failed to get TiDB configuration from cluster [%s/%s], ordinal: %d, error: %v", tc.Namespace, tc.Name, i, err)
-			return false
-		}
-		if clusterInfo.TiDBTokenLimit > 0 && uint(clusterInfo.TiDBTokenLimit) != config.TokenLimit {
-			log.Logf("check [%s/%s] TiDB instance [%d] configuration updated failed: desired [%d], actual [%d] not equal",
-				tc.Namespace, tc.Name, i, clusterInfo.TiDBTokenLimit, config.TokenLimit)
-			return false
-		}
-	}
-	return true
-}
-
-func (oa *operatorActions) checkTiKVConfigUpdated(tc *v1alpha1.TidbCluster, clusterInfo *TidbClusterConfig) bool {
-	// TODO: check if TiKV configuration updated
-	return true
-}
-
-func (oa *operatorActions) checkPrometheus(clusterInfo *TidbClusterConfig) error {
-	ns := clusterInfo.Namespace
-	tcName := clusterInfo.ClusterName
-	return checkPrometheusCommon(tcName, ns, oa.fw)
-}
-
-func (oa *operatorActions) checkGrafanaData(clusterInfo *TidbClusterConfig) error {
-	ns := clusterInfo.Namespace
-	tcName := clusterInfo.ClusterName
-	grafanaClient, err := checkGrafanaDataCommon(tcName, ns, clusterInfo.GrafanaClient, oa.fw)
-	if err != nil {
-		return err
-	}
-	if clusterInfo.GrafanaClient == nil && grafanaClient != nil {
-		clusterInfo.GrafanaClient = grafanaClient
-	}
-	return nil
-}
-
 func getDatasourceID(addr string) (int, error) {
 	u := fmt.Sprintf("http://%s/api/datasources", addr)
 	req, err := http.NewRequest(http.MethodGet, u, nil)
@@ -2641,14 +1436,6 @@ func (oa *operatorActions) CheckRestore(from *TidbClusterConfig, to *TidbCluster
 		return fmt.Errorf("failed to launch restore job: %v", err)
 	}
 	return nil
-}
-
-func (oa *operatorActions) ForceDeploy(info *TidbClusterConfig) error {
-	if err := oa.CleanTidbCluster(info); err != nil {
-		return err
-	}
-
-	return oa.DeployTidbCluster(info)
 }
 
 func (oa *operatorActions) DataIsTheSameAs(tc, otherInfo *TidbClusterConfig) (bool, error) {
@@ -3656,62 +2443,6 @@ func (oa *operatorActions) pumpMembersReadyFn(tc *v1alpha1.TidbCluster) (bool, e
 		}
 	}
 	return true, nil
-}
-
-// FIXME: this duplicates with WaitForTidbClusterReady in crd_test_utils.go, and all functions in it
-func (oa *operatorActions) WaitForTidbClusterReady(tc *v1alpha1.TidbCluster, timeout, pollInterval time.Duration) error {
-	if tc == nil {
-		return fmt.Errorf("tidbcluster is nil, cannot call WaitForTidbClusterReady")
-	}
-	return wait.PollImmediate(pollInterval, timeout, func() (bool, error) {
-		var local *v1alpha1.TidbCluster
-		var err error
-		if local, err = oa.cli.PingcapV1alpha1().TidbClusters(tc.Namespace).Get(tc.Name, metav1.GetOptions{}); err != nil {
-			log.Logf("failed to get tidbcluster: %s/%s, %v", tc.Namespace, tc.Name, err)
-			return false, nil
-		}
-
-		if b, err := oa.pdMembersReadyFn(local); !b && err == nil {
-			log.Logf("pd members are not ready for tc %q", tc.Name)
-			return false, nil
-		}
-		log.Logf("pd members are ready for tc %q", tc.Name)
-
-		if b, err := oa.tikvMembersReadyFn(local); !b && err == nil {
-			log.Logf("tikv members are not ready for tc %q", tc.Name)
-			return false, nil
-		}
-		log.Logf("tikv members are ready for tc %q", tc.Name)
-
-		if b, err := oa.tidbMembersReadyFn(local); !b && err == nil {
-			log.Logf("tidb members are not ready for tc %q", tc.Name)
-			return false, nil
-		}
-		log.Logf("tidb members are ready for tc %q", tc.Name)
-
-		if tc.Spec.TiFlash != nil && tc.Spec.TiFlash.Replicas > int32(0) {
-			if b, err := oa.tiflashMembersReadyFn(local); !b && err == nil {
-				log.Logf("tiflash members are not ready for tc %q", tc.Name)
-				return false, nil
-			}
-			log.Logf("tiflash members are ready for tc %q", tc.Name)
-		} else {
-			log.Logf("no tiflash in tc spec")
-		}
-
-		if tc.Spec.Pump != nil {
-			if b, err := oa.pumpMembersReadyFn(local); !b && err == nil {
-				log.Logf("pump members are not ready for tc %q", tc.Name)
-				return false, nil
-			}
-			log.Logf("pump members are ready for tc %q", tc.Name)
-		} else {
-			log.Logf("no pump in tc spec")
-		}
-
-		log.Logf("TidbCluster is ready")
-		return true, nil
-	})
 }
 
 var dummyCancel = func() {}
